[rebase] seekingmemberpost 경로 변경
