+++ project-service/src/main/java/org/omoknoone/omokhub/project/command/repository/SeekingMemberPostRepository.java
@@ -1,6 +1,6 @@
-package org.omoknoone.omokhub.seekingmemberpost.command.repository;
+package org.omoknoone.omokhub.project.command.repository;
 
-import org.omoknoone.omokhub.seekingmemberpost.command.aggregate.SeekingMemberPost;
+import org.omoknoone.omokhub.project.command.aggregate.SeekingMemberPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SeekingMemberPostRepository extends JpaRepository<SeekingMemberPost, Integer> {
