+++ project-service/src/main/java/org/omoknoone/omokhub/project/command/service/SeekingMemberPostService.java
@@ -1,6 +1,6 @@
-package org.omoknoone.omokhub.seekingmemberpost.command.service;
+package org.omoknoone.omokhub.project.command.service;
 
-import org.omoknoone.omokhub.seekingmemberpost.command.dto.SeekingMemberPostDTO;
+import org.omoknoone.omokhub.project.command.dto.SeekingMemberPostDTO;
 
 public interface SeekingMemberPostService {
     int newPost(SeekingMemberPostDTO seekingMemberPostDTO);
