+++ project-service/src/main/java/org/omoknoone/omokhub/project/query/dto/SeekingMemberPostCriteriaDTO.java
@@ -1,4 +1,4 @@
-package org.omoknoone.omokhub.seekingmemberpost.query.dto;
+package org.omoknoone.omokhub.project.query.dto;
 
 public class SeekingMemberPostCriteriaDTO {
     private String startDate;
