+++ project-service/src/main/java/org/omoknoone/omokhub/project/query/service/SeekingMemberPostServiceImpl.java
@@ -1,9 +1,9 @@
-package org.omoknoone.omokhub.seekingmemberpost.query.service;
+package org.omoknoone.omokhub.project.query.service;
 
-import org.omoknoone.omokhub.seekingmemberpost.query.dto.SeekingMemberPostCriteriaDTO;
-import org.omoknoone.omokhub.seekingmemberpost.query.dto.SeekingMemberPostDTO;
-import org.omoknoone.omokhub.seekingmemberpost.query.entity.SeekingMemberPost;
-import org.omoknoone.omokhub.seekingmemberpost.query.repository.SeekingMemberPostMapper;
+import org.omoknoone.omokhub.project.query.dto.SeekingMemberPostCriteriaDTO;
+import org.omoknoone.omokhub.project.query.dto.SeekingMemberPostDTO;
+import org.omoknoone.omokhub.project.query.entity.SeekingMemberPost;
+import org.omoknoone.omokhub.project.query.repository.SeekingMemberPostMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
